Use cl-labels instead of cl-letf*
As mentioned in info node `(cl) Function Bindings' each PLACE that is
set with cl-letf or cl-letf* must have a well-defined value outside
the body of the macro.  Here this isn't the case, and while it appears
that it works anyway, there is no guarantee it will stay that way and
the byte-compiler also complains about it; so use cl-labels instead.